# docs/openapi.json
{
  "components": {
    "schemas": {
      "AnalyzeOutcome": {
        "discriminator": {
          "mapping": {
            "analyzed": "#/components/schemas/AnalyzeResponse",
            "numeric_breakdown": "#/components/schemas/NumericalFailure"
          },
          "propertyName": "status"
        },
        "oneOf": [
          {
            "$ref": "#/components/schemas/AnalyzeResponse"
          },
          {
            "$ref": "#/components/schemas/NumericalFailure"
          }
        ]
      },
      "AnalyzeRequest": {
        "additionalProperties": false,
        "properties": {
          "arithmetic_mode": {
            "default": "float64",
            "enum": [
              "float64",
              "exact"
            ],
            "title": "Arithmetic Mode",
            "type": "string"
          },
          "system": {
            "$ref": "#/components/schemas/SystemInput"
          }
        },
        "required": [
          "system"
        ],
        "title": "AnalyzeRequest",
        "type": "object"
      },
      "AnalyzeResponse": {
        "additionalProperties": false,
        "properties": {
          "classification": {
            "$ref": "#/components/schemas/ClassificationResult"
          },
          "conditioning": {
            "$ref": "#/components/schemas/ConditionDiagnostic"
          },
          "conditioning_arithmetic_mode": {
            "const": "float64",
            "default": "float64",
            "title": "Conditioning Arithmetic Mode",
            "type": "string"
          },
          "dominance_permutation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RowPermutation"
              },
              {
                "type": "null"
              }
            ]
          },
          "is_square": {
            "title": "Is Square",
            "type": "boolean"
          },
          "methods": {
            "items": {
              "$ref": "#/components/schemas/MethodEligibility"
            },
            "title": "Methods",
            "type": "array"
          },
          "nonzero_permutation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RowPermutation"
              },
              {
                "type": "null"
              }
            ]
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          },
          "shape": {
            "$ref": "#/components/schemas/SystemShape"
          },
          "status": {
            "const": "analyzed",
            "default": "analyzed",
            "title": "Status",
            "type": "string"
          },
          "strict_diagonal_dominance": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Strict Diagonal Dominance"
          },
          "warnings": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "status",
          "request_id",
          "shape",
          "is_square",
          "classification",
          "conditioning",
          "conditioning_arithmetic_mode",
          "strict_diagonal_dominance",
          "dominance_permutation",
          "nonzero_permutation",
          "methods",
          "warnings"
        ],
        "title": "AnalyzeResponse",
        "type": "object"
      },
      "ArithmeticMode": {
        "enum": [
          "float64",
          "exact"
        ],
        "title": "ArithmeticMode",
        "type": "string"
      },
      "ClassificationKind": {
        "enum": [
          "unique",
          "infinite",
          "inconsistent"
        ],
        "title": "ClassificationKind",
        "type": "string"
      },
      "ClassificationResult": {
        "additionalProperties": false,
        "properties": {
          "arithmetic_mode": {
            "$ref": "#/components/schemas/ArithmeticMode"
          },
          "classification": {
            "$ref": "#/components/schemas/ClassificationKind"
          },
          "rank_a": {
            "minimum": 0.0,
            "title": "Rank A",
            "type": "integer"
          },
          "rank_augmented": {
            "minimum": 0.0,
            "title": "Rank Augmented",
            "type": "integer"
          },
          "rank_tolerance": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rank Tolerance"
          },
          "shape": {
            "$ref": "#/components/schemas/SystemShape"
          },
          "tolerance_policy": {
            "title": "Tolerance Policy",
            "type": "string"
          }
        },
        "required": [
          "shape",
          "arithmetic_mode",
          "rank_a",
          "rank_augmented",
          "classification",
          "rank_tolerance",
          "tolerance_policy"
        ],
        "title": "ClassificationResult",
        "type": "object"
      },
      "ConditionDiagnostic": {
        "additionalProperties": false,
        "properties": {
          "approximate_digit_loss": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/NonnegativeFloat"
              },
              {
                "type": "null"
              }
            ]
          },
          "condition_number": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/NonnegativeFloat"
              },
              {
                "type": "null"
              }
            ]
          },
          "status": {
            "enum": [
              "finite",
              "singular",
              "unavailable"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "title": "ConditionDiagnostic",
        "type": "object"
      },
...
      "DirectOptions": {
        "additionalProperties": false,
        "properties": {
          "arithmetic_mode": {
            "default": "float64",
            "enum": [
              "float64",
              "exact"
            ],
            "title": "Arithmetic Mode",
            "type": "string"
          }
        },
        "title": "DirectOptions",
        "type": "object"
      },
      "DirectResult": {
        "additionalProperties": false,
        "properties": {
          "arithmetic_mode": {
            "$ref": "#/components/schemas/ArithmeticMode"
          },
          "augmented_pivot_columns": {
            "items": {
              "type": "integer"
            },
            "title": "Augmented Pivot Columns",
            "type": "array"
          },
          "classification": {
            "$ref": "#/components/schemas/ClassificationResult"
          },
          "contradictory_rows": {
            "items": {
              "type": "integer"
            },
            "title": "Contradictory Rows",
            "type": "array"
          },
          "diagnostics": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DirectDiagnostics"
              },
              {
                "type": "null"
              }
            ]
          },
          "form": {
            "enum": [
              "ref",
              "rref"
            ],
            "title": "Form",
            "type": "string"
          },
          "free_columns": {
            "items": {
              "type": "integer"
            },
            "title": "Free Columns",
            "type": "array"
          },
          "matrix": {
            "$ref": "#/components/schemas/NumericMatrix"
          },
          "method": {
            "enum": [
              "gaussian",
              "gauss_jordan"
            ],
            "title": "Method",
            "type": "string"
          },
          "parametric_solution": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ParametricSolution"
              },
              {
                "type": "null"
              }
            ]
          },
          "pivot_columns": {
            "items": {
              "type": "integer"
            },
            "title": "Pivot Columns",
            "type": "array"
          },
          "row_tolerances": {
            "anyOf": [
              {
                "items": {
                  "type": "number"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Row Tolerances"
          },
          "solution": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/NumericVector"
              },
              {
                "type": "null"
              }
            ]
          },
          "trace": {
            "items": {
              "$ref": "#/components/schemas/RowOperation"
            },
            "title": "Trace",
            "type": "array"
          },
          "warnings": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "method",
          "arithmetic_mode",
          "classification",
          "form",
          "matrix",
          "pivot_columns",
          "free_columns",
          "augmented_pivot_columns",
          "solution",
          "parametric_solution",
          "contradictory_rows",
          "row_tolerances",
          "trace",
          "diagnostics"
        ],
        "title": "DirectResult",
        "type": "object"
      },
      "DirectSolveRequest": {
        "additionalProperties": false,
        "properties": {
          "display": {
            "$ref": "#/components/schemas/DisplayPreferences"
          },
          "method": {
            "enum": [
              "gaussian",
              "gauss_jordan"
            ],
            "title": "Method",
            "type": "string"
          },
          "options": {
            "$ref": "#/components/schemas/DirectOptions"
          },
          "system": {
            "$ref": "#/components/schemas/SystemInput"
          }
        },
        "required": [
          "system",
          "method"
        ],
        "title": "DirectSolveRequest",
        "type": "object"
      },
      "DisplayPreferences": {
        "additionalProperties": false,
        "properties": {
          "decimal_places": {
            "default": 6,
            "maximum": 12.0,
            "minimum": 0.0,
            "title": "Decimal Places",
            "type": "integer"
          },
          "mode": {
            "default": "decimal",
            "enum": [
              "decimal",
              "fraction"
            ],
            "title": "Mode",
            "type": "string"
          }
        },
        "title": "DisplayPreferences",
        "type": "object"
      },
      "ErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "details": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/Issue"
            },
            "title": "Details",
            "type": "array"
          },
          "error": {
            "$ref": "#/components/schemas/Issue"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          },
          "status": {
            "const": "error",
            "default": "error",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "request_id",
          "error",
          "details"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "FloatScalar": {
        "type": "number"
      },
...
      "InputRow": {
        "items": {
          "$ref": "#/components/schemas/NumericToken"
        },
        "maxItems": 12,
        "minItems": 1,
        "type": "array"
      },
      "Issue": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "location": {
            "default": [],
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "location"
        ],
        "title": "Issue",
        "type": "object"
      },
...
      "IterativeOptions": {
        "additionalProperties": false,
        "properties": {
          "auto_reorder_for_diagonal_dominance": {
            "default": true,
            "title": "Auto Reorder For Diagonal Dominance",
            "type": "boolean"
          },
          "auto_reorder_for_nonzero_diagonal": {
            "default": false,
            "title": "Auto Reorder For Nonzero Diagonal",
            "type": "boolean"
          },
          "initial_guess": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/NumericToken"
                },
                "maxItems": 12,
                "minItems": 1,
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Initial Guess"
          },
          "max_iterations": {
            "default": 25,
            "maximum": 500.0,
            "minimum": 1.0,
            "title": "Max Iterations",
            "type": "integer"
          },
          "run_despite_convergence_risk": {
            "default": false,
            "title": "Run Despite Convergence Risk",
            "type": "boolean"
          },
          "tolerance": {
            "default": 1e-08,
            "maximum": 0.01,
            "minimum": 1e-14,
            "title": "Tolerance",
            "type": "number"
          }
        },
        "title": "IterativeOptions",
        "type": "object"
      },
...
      "IterativeSolveRequest": {
        "additionalProperties": false,
        "properties": {
          "display": {
            "$ref": "#/components/schemas/DisplayPreferences"
          },
          "method": {
            "enum": [
              "jacobi",
              "gauss_seidel"
            ],
            "title": "Method",
            "type": "string"
          },
          "options": {
            "$ref": "#/components/schemas/IterativeOptions"
          },
          "system": {
            "$ref": "#/components/schemas/SystemInput"
          }
        },
        "required": [
          "system",
          "method"
        ],
        "title": "IterativeSolveRequest",
        "type": "object"
      },
      "Method": {
        "enum": [
          "gaussian",
          "gauss_jordan",
          "jacobi",
          "gauss_seidel"
        ],
        "type": "string"
      },
      "MethodEligibility": {
        "additionalProperties": false,
        "description": "Eligible means available subject to the stated reordering and risk requirements.",
        "properties": {
          "code": {
            "enum": [
              "available",
              "requires_square",
              "requires_unique",
              "requires_float64",
              "unsafe_diagonal",
              "requires_row_reordering",
              "convergence_risk",
              "numeric_breakdown"
            ],
            "title": "Code",
            "type": "string"
          },
          "convergence": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ConvergenceDiagnostics"
              },
              {
                "type": "null"
              }
            ]
          },
          "eligible": {
            "title": "Eligible",
            "type": "boolean"
          },
          "method": {
            "$ref": "#/components/schemas/Method"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          },
          "requires_risk_override": {
            "default": false,
            "title": "Requires Risk Override",
            "type": "boolean"
          },
          "requires_row_reordering": {
            "default": false,
            "title": "Requires Row Reordering",
            "type": "boolean"
          },
          "row_permutation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RowPermutation"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "method",
          "eligible",
          "code",
          "reason",
          "requires_row_reordering",
          "requires_risk_override",
          "row_permutation",
          "convergence"
        ],
        "title": "MethodEligibility",
        "type": "object"
      },
      "NonnegativeFloat": {
        "$ref": "#/components/schemas/FloatScalar",
        "ge": 0
      },
      "NumericMatrix": {
        "items": {
          "$ref": "#/components/schemas/NumericVector"
        },
        "type": "array"
      },
      "NumericToken": {
        "maxLength": 48,
        "minLength": 1,
        "type": "string"
      },
...
      "NumericVector": {
        "items": {
          "$ref": "#/components/schemas/NumericValue"
        },
        "type": "array"
      },
      "NumericalFailure": {
        "additionalProperties": false,
        "description": "A controlled numerical limitation, without an invented classification or solution.",
        "properties": {
          "arithmetic_mode": {
            "$ref": "#/components/schemas/ArithmeticMode"
          },
          "error": {
            "$ref": "#/components/schemas/Issue"
          },
          "method": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Method"
              },
              {
                "type": "null"
              }
            ]
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          },
          "shape": {
            "$ref": "#/components/schemas/SystemShape"
          },
          "status": {
            "const": "numeric_breakdown",
            "default": "numeric_breakdown",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "request_id",
          "method",
          "arithmetic_mode",
          "shape",
          "error"
        ],
        "title": "NumericalFailure",
        "type": "object"
      },
...
      "ProblemMetadata": {
        "additionalProperties": false,
        "properties": {
          "arithmetic_mode": {
            "$ref": "#/components/schemas/ArithmeticMode"
          },
          "is_square": {
            "title": "Is Square",
            "type": "boolean"
          },
          "original_system": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FloatSystem"
              },
              {
                "$ref": "#/components/schemas/ExactSystem"
              }
            ],
            "title": "Original System"
          },
          "shape": {
            "$ref": "#/components/schemas/SystemShape"
          }
        },
        "required": [
          "shape",
          "is_square",
          "arithmetic_mode",
          "original_system"
        ],
        "title": "ProblemMetadata",
        "type": "object"
      },
...
      "ReportMetadata": {
        "additionalProperties": false,
        "description": "Use problem + result + this metadata; traces are never duplicated.",
        "properties": {
          "display": {
            "$ref": "#/components/schemas/DisplayPreferences"
          },
          "fraction_values": {
            "enum": [
              "exact",
              "approximate"
            ],
            "title": "Fraction Values",
            "type": "string"
          },
          "method": {
            "$ref": "#/components/schemas/Method"
          },
          "options": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DirectOptions"
              },
              {
                "$ref": "#/components/schemas/IterativeOptions"
              }
            ],
            "title": "Options"
          },
          "schema_version": {
            "const": 1,
            "default": 1,
            "title": "Schema Version",
            "type": "integer"
          },
          "solver_version": {
            "const": "0.1.0",
            "default": "0.1.0",
            "title": "Solver Version",
            "type": "string"
          }
        },
        "required": [
          "schema_version",
          "solver_version",
          "method",
          "options",
          "display",
          "fraction_values"
        ],
        "title": "ReportMetadata",
        "type": "object"
      },
...
      "RowPermutation": {
        "additionalProperties": false,
        "description": "order[j] is the original equation moved to row j; variables never move.",
        "properties": {
          "order": {
            "items": {
              "type": "integer"
            },
            "maxItems": 12,
            "minItems": 1,
            "title": "Order",
            "type": "array"
          },
          "purpose": {
            "enum": [
              "identity",
              "strict_diagonal_dominance",
              "nonzero_diagonal"
            ],
            "title": "Purpose",
            "type": "string"
          }
        },
        "required": [
          "order",
          "purpose"
        ],
        "title": "RowPermutation",
        "type": "object"
      },
      "SolveOutcome": {
        "discriminator": {
          "mapping": {
            "completed": "#/components/schemas/SolveResponse",
            "numeric_breakdown": "#/components/schemas/NumericalFailure"
          },
          "propertyName": "status"
        },
        "oneOf": [
          {
            "$ref": "#/components/schemas/SolveResponse"
          },
          {
            "$ref": "#/components/schemas/NumericalFailure"
          }
        ]
      },
      "SolveResponse": {
        "additionalProperties": false,
        "description": "Completed means the computation returned a result, not that it converged.\n\nInspect result.classification for direct methods and result.status for iteration.\nConditioning always estimates the original coefficient matrix in float64.",
        "properties": {
          "conditioning": {
            "$ref": "#/components/schemas/ConditionDiagnostic"
          },
          "conditioning_arithmetic_mode": {
            "const": "float64",
            "default": "float64",
            "title": "Conditioning Arithmetic Mode",
            "type": "string"
          },
          "problem": {
            "$ref": "#/components/schemas/ProblemMetadata"
          },
          "report": {
            "$ref": "#/components/schemas/ReportMetadata"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          },
          "result": {
            "discriminator": {
              "mapping": {
                "gauss_jordan": "#/components/schemas/DirectResult",
                "gauss_seidel": "#/components/schemas/IterativeResult",
                "gaussian": "#/components/schemas/DirectResult",
                "jacobi": "#/components/schemas/IterativeResult"
              },
              "propertyName": "method"
            },
            "oneOf": [
              {
                "$ref": "#/components/schemas/DirectResult"
              },
              {
                "$ref": "#/components/schemas/IterativeResult"
              }
            ],
            "title": "Result"
          },
          "status": {
            "const": "completed",
            "default": "completed",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "request_id",
          "problem",
          "result",
          "conditioning",
          "conditioning_arithmetic_mode",
          "report"
        ],
        "title": "SolveResponse",
        "type": "object"
      },
...
      "SystemInput": {
        "additionalProperties": false,
        "properties": {
          "a": {
            "items": {
              "$ref": "#/components/schemas/InputRow"
            },
            "maxItems": 12,
            "minItems": 1,
            "title": "A",
            "type": "array"
          },
          "b": {
            "items": {
              "$ref": "#/components/schemas/NumericToken"
            },
            "maxItems": 12,
            "minItems": 1,
            "title": "B",
            "type": "array"
          }
        },
        "required": [
          "a",
          "b"
        ],
        "title": "SystemInput",
        "type": "object"
      },
      "SystemShape": {
        "additionalProperties": false,
        "properties": {
          "equations": {
            "maximum": 12.0,
            "minimum": 1.0,
            "title": "Equations",
            "type": "integer"
          },
          "unknowns": {
            "maximum": 12.0,
            "minimum": 1.0,
            "title": "Unknowns",
            "type": "integer"
          }
        },
        "required": [
          "equations",
          "unknowns"
        ],
        "title": "SystemShape",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "TULYA API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/analyze": {
      "post": {
        "description": "Classify a system and explain conditional eligibility for all four methods.",
        "operationId": "analyze_system",
        "parameters": [
          {
            "description": "Optional correlation ID (1–64 ASCII letters, digits, '-' or '_'). Invalid values are replaced; the final ID is returned in X-Request-ID.",
            "in": "header",
            "name": "x-request-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional correlation ID (1–64 ASCII letters, digits, '-' or '_'). Invalid values are replaced; the final ID is returned in X-Request-ID.",
              "title": "X-Request-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalyzeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalyzeOutcome"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation ID.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid input or resource-policy violation.",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation ID.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unexpected internal error with correlation ID.",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation ID.",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "summary": "Analyze System",
        "tags": [
          "solver"
        ]
      }
    },
    "/api/v1/solve": {
      "post": {
        "description": "Return full results, including mathematical non-success outcomes at HTTP 200.",
        "operationId": "solve_system",
        "parameters": [
          {
            "description": "Optional correlation ID (1–64 ASCII letters, digits, '-' or '_'). Invalid values are replaced; the final ID is returned in X-Request-ID.",
            "in": "header",
            "name": "x-request-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional correlation ID (1–64 ASCII letters, digits, '-' or '_'). Invalid values are replaced; the final ID is returned in X-Request-ID.",
              "title": "X-Request-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "discriminator": {
                  "mapping": {
                    "gauss_jordan": "#/components/schemas/DirectSolveRequest",
                    "gauss_seidel": "#/components/schemas/IterativeSolveRequest",
                    "gaussian": "#/components/schemas/DirectSolveRequest",
                    "jacobi": "#/components/schemas/IterativeSolveRequest"
                  },
                  "propertyName": "method"
                },
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/DirectSolveRequest"
                  },
                  {
                    "$ref": "#/components/schemas/IterativeSolveRequest"
                  }
                ],
                "title": "Payload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SolveOutcome"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation ID.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid input or resource-policy violation.",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation ID.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unexpected internal error with correlation ID.",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation ID.",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "summary": "Solve System",
        "tags": [
          "solver"
        ]
      }
    }
  }
}